+++ Section27/CitiesManager/CitiesManager.WebAPI/Controllers/CustomControllerBase.cs
@@ -2,7 +2,7 @@
 
 namespace CitiesManager.WebAPI.Controllers
 {
- [Route("api/[controller]")]
+ [Route("api/v{version:apiVersion}/[controller]")]
  [ApiController]
  public class CustomControllerBase : ControllerBase
  {
