+++ Section27/CitiesManager/CitiesManager.WebAPI/Controllers/v2/CitiesController.cs
@@ -1,0 +1,33 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using CitiesManager.WebAPI.DatabaseContext;
+using CitiesManager.WebAPI.Models;
+
+namespace CitiesManager.WebAPI.Controllers.v2
+{
+    [ApiVersion("2.0")]
+    public class CitiesController : CustomControllerBase
+    {
+        private readonly ApplicationDbContext _context;
+
+        public CitiesController(ApplicationDbContext context)
+        {
+            _context = context;
+        }
+
+        // GET: api/Cities
+        /// <summary>
+        /// To get list of  cities (only city name) from 'cities' table
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet]
+        //[Produces("application/xml")]
+        public async Task<ActionResult<IEnumerable<string>>> GetCities()
+        {
+            var cities = await _context.Cities
+             .OrderBy(temp => temp.CityName).Select(t=>t.CityName).ToListAsync();
+            return cities;
+        }
+
+    }
+}
