+++ Section27/CitiesManager/CitiesManager.WebAPI/Program.cs
@@ -1,5 +1,6 @@
 using CitiesManager.WebAPI.DatabaseContext;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Versioning;
 using Microsoft.EntityFrameworkCore;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -13,7 +14,10 @@ builder.Services.AddControllers(options =>
 })
     .AddXmlSerializerFormatters();
 
-
+builder.Services.AddApiVersioning(config =>
+{
+    config.ApiVersionReader = new UrlSegmentApiVersionReader();
+});
 
 builder.Services.AddDbContext<ApplicationDbContext>(options =>
 {
