+++ Section27/CitiesManager/CitiesManager.WebAPI/api.xml
@@ -4,17 +4,23 @@
         <name>CitiesManager.WebAPI</name>
     </assembly>
     <members>
-        <member name="M:CitiesManager.WebAPI.Controllers.CitiesController.GetCities">
+        <member name="M:CitiesManager.WebAPI.Controllers.v1.CitiesController.GetCities">
             <summary>
             To get list of  cities (including city id and city name) from 'cities' table
             </summary>
             <returns></returns>
         </member>
-        <member name="M:CitiesManager.WebAPI.Controllers.CitiesController.GetCity(System.Guid)">
+        <member name="M:CitiesManager.WebAPI.Controllers.v1.CitiesController.GetCity(System.Guid)">
             <summary>
             GET: api/Cities/5
             </summary>
             <param name="cityID"></param>
+            <returns></returns>
+        </member>
+        <member name="M:CitiesManager.WebAPI.Controllers.v2.CitiesController.GetCities">
+            <summary>
+            To get list of  cities (only city name) from 'cities' table
+            </summary>
             <returns></returns>
         </member>
         <member name="T:CitiesManager.WebAPI.Migrations.Initial">
